fix(preview): upgrade monaco-editor to 0.56.0 with reorganized entry points
intent(monaco): renovate の 0.56.0 更新が lockfile 生成に失敗し CI 全 job が落ちていた状態を解消する
constraint(monaco): 0.56.0 の exports map は `./*` を `./esm/vs/*.js` へ写すため、`monaco-editor/esm/vs/...` の deep import は全て解決不能になる。実ファイルは同じ場所にあり指定子から `esm/vs/` prefix を落とすだけで解決する
constraint(monaco-patch): patch は配布物 dom.js の export 行の内容に依存し、0.56.0 で ConnectionObserverElement / getExternalFocusWindow が増えたため 0.55.1 用の hunk は適用できない
decision(monaco-patch): patch key をバージョン限定のまま 0.56.0 へ更新し、patch ファイル名も pnpm 生成名 (monaco-editor@0.56.0.patch) に合わせる。型宣言側からの参照は version 非依存の表記にして SSOT を pnpm-workspace.yaml 側へ寄せる
rejected(monaco): 0.56.0 新設の getExternalFocusWindow で patch を廃止する案 — 登録関数 registerExternalFocusChecker が registerWindow と同じく standalone 配布の tree-shake で落ちており、配布物から登録できない
rejected(monaco): 依存更新と移行を別 commit に分ける案 — package.json だけ進んだ中間 commit は pnpm install が ERR_PNPM_UNUSED_PATCH で失敗し、bisect 不能な壊れた状態になる
learned(monaco): registerWindow は 0.56.0 でも IIFE 内に実体が残り export リストからのみ落ちる。0.56.0 は tree-shake 可能な部分 import (monaco-editor/editor + features/*/register) をパッケージ内に持ち、typescriptDefaults は型付き export になった
