+++ apps/renderer/package.json
@@ -32,7 +32,7 @@
     "marked": "18.0.7",
     "material-icon-theme": "5.37.0",
     "mermaid": "11.16.0",
-    "monaco-editor": "0.55.1",
+    "monaco-editor": "0.56.0",
     "pinia": "4.0.2",
     "shiki": "4.3.1",
     "vue": "3.5.40"
