+++ pnpm-workspace.yaml
@@ -55,6 +55,6 @@ minimumReleaseAgeExclude:
 shellEmulator: true
 
 patchedDependencies:
-  # バージョン限定: patch は 0.55.1 の配布物の行位置に依存する。name-only key だと
+  # バージョン限定: patch は 0.56.0 の配布物の行位置に依存する。name-only key だと
   # 将来バージョンにも適用され、失敗理由が patch 側の腐敗として現れて分かりにくい
-  "monaco-editor@0.55.1": patches/monaco-editor.patch
+  monaco-editor@0.56.0: patches/monaco-editor@0.56.0.patch
